requirements.txt
python-telegram-bot
transformers
torch
python-dotenv
huggingface-hub
accelerate
nest-asyncio
